+++ src/main/java/koreatech/in/service/admin/AdminUserService.java
@@ -10,9 +10,8 @@ import koreatech.in.dto.admin.user.request.LoginRequest;
 import koreatech.in.dto.admin.user.request.NewOwnersCondition;
 import koreatech.in.dto.admin.user.response.LoginResponse;
 import koreatech.in.dto.admin.user.response.NewOwnersResponse;
-import koreatech.in.dto.admin.user.student.StudentResponse;
+import koreatech.in.dto.admin.user.student.response.StudentResponse;
 import koreatech.in.dto.admin.user.response.OwnerResponse;
-import koreatech.in.dto.normal.user.request.UpdateUserRequest;
 import koreatech.in.dto.normal.user.student.request.StudentUpdateRequest;
 
 public interface AdminUserService {
